% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoscaling_df.R
\name{autoscaling_df}
\alias{autoscaling_df}
\title{Autoscaling}
\usage{
autoscaling(df)
}
\arguments{
\item{df}{data frame with the last column corresponding to the class or group being used.}
}
\description{
Function for data pre-treatment.
}
\examples{
data(nir_seed)
df<-autoscaling_df(nir_seed)
a<-plotmean_df(nir_seed)
b<-plotmean_df(df)
library(ggpubr)
ggarrange(a, b,
labels = c("A", "B"),
ncol = 2, nrow = 1)
}
