% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center.R
\name{center_df}
\alias{center_df}
\title{Center}
\usage{
scale_center(dados)
}
\arguments{
\item{dados}{data frame with the last column corresponding to the class or group being used.}
}
\description{
Function for data pre-treatment.
}
\examples{
data(nir_seed)
dados<-center_df(nir_seed)
a<-vizplotmean(nir_seed)
b<-vizplotmean(dados)
library(ggpubr)
ggarrange(a, b,
labels = c("A", "B"),
ncol = 2, nrow = 1)
}
