% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_df.R
\name{classification_df}
\alias{classification_df}
\title{classification_df}
\usage{
classification_df(df)
}
\arguments{
\item{df}{data frame with the last column corresponding to the class or group being used.}

\item{splitting}{Establishes the proportion and creates divisions balanced the training data and test.}

\item{algorithm}{Algorithm that will be used to create the models.}

\item{method_cv}{The resampling method: "boot", "boot632", "optimism_boot", "boot_all", "cv", "repeatedcv", "LOOCV", "LGOCV"}

\item{number_cv}{Either the number of folds or number of resampling iterations}

\item{repeat_cv}{For repeated k-fold cross-validation only: the number of complete sets of folds to compute}
}
\description{
Function to develop classification models
}
\examples{
data(nir_seed)
dados<-classification_df(nir_seed)
}
