% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/der_SG.R
\name{der_SG_df}
\alias{der_SG_df}
\title{Applying derivatives with Savitzky-Golay smoothing}
\usage{
der_SG(df)
}
\arguments{
\item{df}{data frame with the last column corresponding to the class or group being used.}

\item{m}{the differentiation order.}

\item{p}{the polynomial order.}

\item{w}{a window size (must be odd).}
}
\description{
Function for data pre-treatment.
}
\examples{
data(nir_seed)
a<-vizplotmean(nir_seed)

#Applying 1st derivative with Savitzky-Golay smoothing.
df<-der_SG_df(nir_seed,1,2,11)
b<-vizplotmean(df)

#Applying 2nd derivative with Savitzky-Golay smoothing.
df<-der_SG_df(nir_seed,2,2,11)
c<-vizplotmean(df)
library(ggpubr)
ggarrange(a, b, c, 
labels = c("A", "B", "C"),
ncol = 2, nrow = 2)
}
