% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_classifier_df.R
\name{load_classifier_df}
\alias{load_classifier_df}
\title{load_classifier_df}
\usage{
load_classifier_df(df, model="model.RData")
}
\arguments{
\item{df}{data frame with the last column corresponding to the class or group being used.}

\item{model}{classifier model name.}
}
\description{
function to load and apply classifier created with the classification_df function
}
