% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msc.R
\name{msc}
\alias{msc}
\title{Applying Multiplicative Scatter Correction}
\usage{
msc(dados)
}
\arguments{
\item{dados}{data frame with the last column corresponding to the class or group being used.}
}
\description{
Function for data pre-treatment.
}
