% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_df.R
\name{pca_df}
\alias{pca_df}
\title{pca_df}
\usage{
pca_df()
}
\arguments{
\item{df}{data frame}

\item{compx}{numero especificando qual o componente sera disposto no eixo y}
}
\description{
funcao para analise exploratoria de componentes principais
}
\examples{
data(nir_seed)
pca_df(nir_seed)
}
