% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmean_df.R
\name{plotmean_df}
\alias{plotmean_df}
\title{plotmean_df}
\usage{
plotmean_df(df)
}
\arguments{
\item{df}{data frame}
}
\description{
funcao para plotagens de graficos do espectro espectro medio das classes
}
\details{
O nome das variaveis do espectro deve ser um numero. Chame os arquivos com check.name = F
}
\examples{
data(nir_seed)
plotmean_df(nir_seed)
}
