% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizplorraw.R
\name{plotraw_df}
\alias{plotraw_df}
\title{Visualization of mean spectra per class}
\usage{
plotraw_df(df)
}
\arguments{
\item{df}{data frame with the last column corresponding to the class or group being used.}
}
\description{
Function to create graphics with NIR spectra
}
\details{
Call files with check.name = T. The name of the spectrum variables must be numeric.
}
\examples{
data(nir_seed)
plotraw_df(nir_seed)
}
