% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_df.R
\name{prediction_df}
\alias{prediction_df}
\title{prediction_df}
\usage{
prediction_df(df, model="model.RData")
}
\arguments{
\item{df}{data frame}

\item{model}{nome do modelo que se deseja carregar}
}
\description{
funcao para aplicar o classificador em novos dados sem rotulos
}
