% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snv.R
\name{snv}
\alias{snv}
\title{Applying standard normal variate}
\usage{
snv(dados)
}
\arguments{
\item{dados}{data frame with the last column corresponding to the class or group being used.}
}
\description{
Function for data pre-treatment.
}
