% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snv.R
\name{snv_df}
\alias{snv_df}
\title{Applying standard normal variate}
\usage{
snv(df)
}
\arguments{
\item{df}{data frame with the last column corresponding to the class or group being used.}
}
\description{
Function for data pre-treatment.
}
