% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizplorraw.R
\name{vizplotraw}
\alias{vizplotraw}
\title{Visualization of mean spectra per class}
\usage{
vizplotraw(dados)
}
\arguments{
\item{dados}{data frame with the last column corresponding to the class or group being used.}
}
\description{
Function to create graphics with NIR spectra
}
\details{
Call files with check.name = T. The name of the spectrum variables must be numeric.
}
\examples{
data(nir_seed)
vizplotraw(nir_seed)
}
