% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/der_SG.R
\name{der_SG}
\alias{der_SG}
\title{der_SG Applying derivatives with smoothing}
\usage{
der_SG( )
}
\arguments{
\item{df}{data frame}

\item{m}{the differentiation order.}

\item{p}{the polynomial order.}

\item{w}{a window size (must be odd).}
}
\description{
funcao para pre-processamento com a aplicacao de derivadas com a suavizacao de Savitzky-Golay
}
\examples{
data(nir_seed)
a<-plotmean_df(nir_seed)

#Applying 1st derivative with Savitzky-Golay smoothing
df<-der_SG(nir_seed,1,2,11)
b<-plotmean_df(df)

#Applying 2nd derivative with Savitzky-Golay smoothing
df<-der_SG(nir_seed,2,2,11)
c<-plotmean_df(df)
library(ggpubr)
ggarrange(a, b, c, 
labels = c("A", "B", "C"),
ncol = 2, nrow = 2)
}
