% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msc_df.R
\name{msc_df}
\alias{msc_df}
\title{msc_df}
\usage{
msc_df()
}
\arguments{
\item{dados}{data frame with the last column corresponding to the class or group being used.}
}
\description{
funcao para pre-processamento dos espectros com o Multiplicative Scatter Correction
}
