% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca_df}
\alias{pca_df}
\title{Autoscaling and center}
\usage{
scale_center(df)
}
\arguments{
\item{df}{data frame with the last column corresponding to the class or group being used.}

\item{axes}{a numeric vector of length 2 specifying the dimensions to be plotted.}
}
\description{
Function for data pre-treatment.
}
\examples{
data(nir_seed)
pca_df(nir_seed)
}
