% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmean_df.R
\name{plotmean_df}
\alias{plotmean_df}
\title{Visualization of raw spectra}
\usage{
plotmean_df(df)
}
\arguments{
\item{df}{data frame with the last column corresponding to the class or group being used.}
}
\description{
Function to create graphics with NIR spectra
}
\details{
Call files with check.name = T. The name of the spectrum variables must be numeric.
}
\examples{
data(nir_seed)
plotmean_df(nir_seed)
}
