% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotraw_df.R
\name{plotraw_df}
\alias{plotraw_df}
\title{Plotraw_df}
\usage{
plotraw_df(df)
}
\arguments{
\item{df}{data frame}
}
\description{
funcao para plotagens de graficos do espectro inteiro
}
\details{
O nome das variaveis do espectro deve ser um numero. Chame os arquivos com check.name = F
}
\examples{
data(nir_seed)
plotraw_df(nir_seed)
}
